Split vector analysis by fact family
R5.1 of the transform cleanup plan: vector/analysis/mod.rs becomes
conditions.rs (DNF), dependencies.rs (dependency/occurrence
projection), effects.rs (EffectAtom vocabulary + conflict axioms +
decoration), and uses.rs (SignalUseTable aggregation) behind an
unchanged facade that keeps the orchestrator and AnalysisError. Move
only: pub(super) on sibling-crossed items and per-file imports; all
files under the plan threshold.

transform 385, vector_mode 35/35, clippy clean.
