+++ porting/journal/2026-07-19.md
@@ -1,5 +1,29 @@
 # 2026-07-19
 
+## Split vector analysis by fact family (R5.1)
+
+Commit date: 2026-07-19
+Source commit: this commit
+
+`vector/analysis/mod.rs` (1,853 production lines) becomes four fact-family
+files behind an unchanged public facade:
+
+- `conditions.rs` (242) — execution-condition DNF vocabulary and
+  construction (C++ `conditionAnnotation` counterpart);
+- `dependencies.rs` (707) — dependency/occurrence projection and
+  `SignalAnalysisContext` (C++ `getSignalDependencies`/`OccMarkup`);
+- `effects.rs` (405) — `EffectAtom` vocabulary, the `effects_conflict`
+  domain axioms, foreign signatures, effect decoration/propagation;
+- `uses.rs` (352) — `SignalUseTable` occurrence aggregation;
+- `mod.rs` (128) — facade re-exports plus the `VectorSignalAnalysis`
+  orchestrator, `AnalysisError`, and `analyze_vector_signals`.
+
+Move only: bodies unchanged apart from `pub(super)` on sibling-crossed
+helpers/fields and per-file imports (relocated tests re-import what the old
+facade glob provided). All production files are under the plan threshold.
+Validation: transform 385, `vector_mode` oracle 35/35, clippy `-D warnings`
+clean, fmt applied.
+
 ## Split the vector plan verifier by responsibility (R5.2)
 
 Commit date: 2026-07-19
