+++ porting/journal/README.md
@@ -18,8 +18,8 @@ Files are listed chronologically (oldest day first). Inside each file, entries a
 - [2026-02-23](./2026-02-23.md) (26 entries)
 - [2026-02-24](./2026-02-24.md) (23 entries)
 - [2026-02-25](./2026-02-25.md) (38 entries)
-- [2026-02-26](./2026-02-26.md) (17 entries)
-- [2026-02-27](./2026-02-27.md) (17 entries)
+- [2026-02-26](./2026-02-26.md) (18 entries)
+- [2026-02-27](./2026-02-27.md) (18 entries)
 - [2026-02-28](./2026-02-28.md) (1 entry)
 - [2026-03-01](./2026-03-01.md) (5 entries)
 - [2026-03-02](./2026-03-02.md) (2 entries)
@@ -49,7 +49,7 @@ Files are listed chronologically (oldest day first). Inside each file, entries a
 - [2026-03-29](./2026-03-29.md) (6 entries)
 - [2026-03-30](./2026-03-30.md) (6 entries)
 - [2026-03-31](./2026-03-31.md) (1 entry)
-- [2026-04-01](./2026-04-01.md) (17 entries)
+- [2026-04-01](./2026-04-01.md) (18 entries)
 - [2026-04-02](./2026-04-02.md) (1 entry)
 - [2026-04-03](./2026-04-03.md) (8 entries)
 - [2026-04-04](./2026-04-04.md) (3 entries)
@@ -78,10 +78,10 @@ Files are listed chronologically (oldest day first). Inside each file, entries a
 - [2026-05-02](./2026-05-02.md) (3 entries)
 - [2026-05-03](./2026-05-03.md) (5 entries)
 - [2026-05-04](./2026-05-04.md) (8 entries)
-- [2026-05-05](./2026-05-05.md) (17 entries)
+- [2026-05-05](./2026-05-05.md) (18 entries)
 - [2026-05-06](./2026-05-06.md) (22 entries)
 - [2026-05-07](./2026-05-07.md) (7 entries)
-- [2026-05-09](./2026-05-09.md) (17 entries)
+- [2026-05-09](./2026-05-09.md) (18 entries)
 - [2026-05-10](./2026-05-10.md) (1 entry)
 - [2026-05-11](./2026-05-11.md) (1 entry)
 - [2026-05-12](./2026-05-12.md) (7 entries)
@@ -98,7 +98,7 @@ Files are listed chronologically (oldest day first). Inside each file, entries a
 - [2026-06-09](./2026-06-09.md) (20 entries)
 - [2026-06-10](./2026-06-10.md) (6 entries)
 - [2026-06-14](./2026-06-14.md) (1 entry)
-- [2026-06-15](./2026-06-15.md) (17 entries)
+- [2026-06-15](./2026-06-15.md) (18 entries)
 - [2026-06-16](./2026-06-16.md) (1 entry)
 - [2026-06-17](./2026-06-17.md) (2 entries)
 - [2026-06-19](./2026-06-19.md) (1 entry)
@@ -106,16 +106,16 @@ Files are listed chronologically (oldest day first). Inside each file, entries a
 - [2026-06-21](./2026-06-21.md) (6 entries)
 - [2026-06-22](./2026-06-22.md) (2 entries)
 - [2026-07-05](./2026-07-05.md) (5 entries)
-- [2026-07-07](./2026-07-07.md) (17 entries)
+- [2026-07-07](./2026-07-07.md) (18 entries)
 - [2026-07-08](./2026-07-08.md) (1 entry)
 - [2026-07-09](./2026-07-09.md) (3 entries)
 - [2026-07-10](./2026-07-10.md) (6 entries)
 - [2026-07-11](./2026-07-11.md) (1 entry)
 - [2026-07-12](./2026-07-12.md) (6 entries)
-- [2026-07-13](./2026-07-13.md) (17 entries)
+- [2026-07-13](./2026-07-13.md) (18 entries)
 - [2026-07-14](./2026-07-14.md) (5 entries)
 - [2026-07-15](./2026-07-15.md) (1 entry)
 - [2026-07-16](./2026-07-16.md) (1 entry)
 - [2026-07-17](./2026-07-17.md) (20 entries)
 - [2026-07-18](./2026-07-18.md) (5 entries)
-- [2026-07-19](./2026-07-19.md) (17 entries)
+- [2026-07-19](./2026-07-19.md) (18 entries)
